requirements.txt
Django>=3.2.0,<4.0
djangorestframework>=3.12.0
psycopg2>=2.9.9,<3.0